% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{process_sumstats_for_manhattan}
\alias{process_sumstats_for_manhattan}
\title{Process summary statistics for Manhattan plot}
\usage{
process_sumstats_for_manhattan(
  dat,
  chr_col = "chromosome",
  bp_col = "base_pair_location",
  p_col = "p_value"
)
}
\arguments{
\item{dat}{data.table containing summary statistics}

\item{chr_col}{character string containing the column name for chromosome}

\item{bp_col}{character string containing the column name for position}

\item{p_col}{character string containing the column name for p-value}
}
\value{
list containing data.table with updated chromosome and position
columns, axis set, and y-axis limits
}
\description{
Code originally from Daniel Roelfs, modified by Tom Willis
https://danielroelfs.com/blog/how-i-create-manhattan-plots-using-ggplot/
}
\author{
Daniel Roelfs

Tom Willis
}
