% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{scale_x_neglog10}
\alias{scale_x_neglog10}
\title{-log10 transformation of the x-axis}
\usage{
scale_x_neglog10(...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}}
}
\value{
ggplot2::scale_x_continuous() with -log10 transformation
}
\description{
-log10 transformation of the x-axis
}
\author{
Tom Willis
}
