% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_stdev.R
\name{sdY.est}
\alias{sdY.est}
\title{Estimate continuous trait standard deviation given vectors of squared standard errors, MAF estimates, and sample size}
\usage{
sdY.est(vbeta, maf, n)
}
\arguments{
\item{vbeta}{vector of variance of coefficients}

\item{maf}{vector of MAF (same length as vbeta)}

\item{n}{sample size}
}
\value{
estimated standard deviation of Y
}
\description{
Estimate is based on var(beta-hat) = var(Y) / (n * var(X))
var(X) = 2\emph{maf}(1-maf)
so we can estimate var(Y) by regressing n*var(X) against 1/var(beta)
}
\author{
Chris Wallace
}
