% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{update_meta_analysis_estimates}
\alias{update_meta_analysis_estimates}
\title{Update existing effect estimates and standard errors using inverse variance-weighted fixed-effects meta-analysis}
\usage{
update_meta_analysis_estimates(
  dat,
  study_name,
  effect_label = "BETA",
  se_label = "SE",
  z2_label = "Z2",
  p_label = "P",
  wt_label = "wt"
)
}
\arguments{
\item{dat}{data.table containing effect estimates and standard errors}

\item{study_name}{character string containing the name of the study}

\item{effect_label}{character string containing the label for the effect estimate}

\item{se_label}{character string containing the label for the standard error}

\item{z2_label}{character string containing the label for the squared z-score}

\item{p_label}{character string containing the label for the p-value}

\item{wt_label}{character string containing the label for the weight}
}
\value{
data.table with updated effect estimates, standard errors, z-scores, and p-values
}
\description{
Update existing effect estimates and standard errors using inverse variance-weighted fixed-effects meta-analysis
}
\author{
Tom Willis
}
